+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v0.0.7
 
 - fixes an issue that could cause the overlay to close when changing months
+- fixes overlay positioning (adds scroll offsets)
 - adds missing ESM dependencies